Moving breadcrumbs logic into useBreadcrumbsItems hook. Using reatomAsync (#922)
* chore: add delay util function

* feat: breadcurmbs items resource (no proper debounce)

* chore: remove delay function

* refactor: use reactAsync for fetching breadcrumbs data

* chore: move logic from BreadcrumbsPanel into useBreadcrumbsItems hook

* fix: fix aborting requests
