+++ src/features/breadcrumbs/atoms/breadcrumbsItemsAtom.ts
@@ -1,6 +1,40 @@
-import { action, atom } from '@reatom/core';
-import { getBboxForGeometry } from '~utils/map/camera';
+import { action, type AsyncCtx, atom, reatomAsync, withAbort } from '@reatom/framework';
+import { debounce } from '@github/mini-throttle';
 import { setCurrentMapBbox, type Bbox } from '~core/shared_state/currentMapPosition';
+import { getBboxForGeometry } from '~utils/map/camera';
+import { getBoundaries } from '~core/api/boundaries';
+import { isAbortError } from '~core/api_client/errors';
+import { getCenterFromPosition } from '../helpers/breadcrumbsHelpers';
+import type { CurrentMapPositionAtomState } from '~core/shared_state/currentMapPosition';
+
+const debouncedItemsFetch = debounce(
+  async (ctx: AsyncCtx, position: CurrentMapPositionAtomState) => {
+    if (position) {
+      try {
+        const coords: [number, number] = getCenterFromPosition(position);
+        const response = await getBoundaries(coords, ctx.controller);
+
+        if (response) {
+          breadcrumbsItemsAtom(ctx, response.features);
+        } else {
+          breadcrumbsItemsAtom(ctx, null);
+        }
+      } catch (error) {
+        if (!isAbortError(error)) {
+          console.error('Error when trying to retrieve boundaries:', error);
+        }
+      }
+    }
+  },
+  1000,
+);
+
+export const fetchBreadcrumbsItems = reatomAsync(
+  async (ctx, position: CurrentMapPositionAtomState) => {
+    debouncedItemsFetch(ctx, position);
+  },
+  'breadcrumbsItemsResource',
+).pipe(withAbort());
 
 export const breadcrumbsItemsAtom = atom<GeoJSON.Feature[] | null>(
   null,
