+++ src/features/breadcrumbs/hooks/useBreadcrumbsItems.ts
@@ -1,0 +1,19 @@
+import { useAction, useAtom } from '@reatom/npm-react';
+import { useEffect } from 'react';
+import { currentMapPositionAtom } from '~core/shared_state/currentMapPosition';
+import {
+  breadcrumbsItemsAtom,
+  fetchBreadcrumbsItems,
+} from '../atoms/breadcrumbsItemsAtom';
+
+export function useBreadcrumbsItems() {
+  const [items] = useAtom(breadcrumbsItemsAtom);
+  const [currentMapPosition] = useAtom(currentMapPositionAtom);
+  const fetchBreadcrumbs = useAction(fetchBreadcrumbsItems);
+
+  useEffect(() => {
+    fetchBreadcrumbs(currentMapPosition);
+  }, [currentMapPosition, fetchBreadcrumbs]);
+
+  return items;
+}
